requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
sentence-transformers>=2.2.0
spacy>=3.7.2
xgboost>=1.5.0
shap>=0.41.0
lime>=0.2.0
google-generativeai

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.3.0

# Document Processing
PyMuPDF>=1.26.1
python-docx>=1.2.0
beautifulsoup4>=4.9.3
python-magic>=0.4.27
chardet>=4.0.0

# Web and API
requests>=2.26.0
schedule>=1.2.2
openai>=1.0.0

# Environment and Utils
python-dotenv>=0.19.0
tqdm>=4.65.0

# Development
pytest>=6.2.5
black>=21.7b0
flake8>=3.9.0 